Refactored utils module into facade and entity-consistency-checker modules.
