+++ bean-consistency-checker/README.md
@@ -1,0 +1,3 @@
+# Lightblue-entity-consistency-checker
+
+A utility to compare POJOs field by field. Use it during migration to lightblue to ensure consistency between entities returned by legacy and lightblue DAOs.
