+++ bean-consistency-checker/src/test/java/com/redhat/lightblue/migrator/consistency/CountryInCountry.java
@@ -1,0 +1,21 @@
+package com.redhat.lightblue.migrator.consistency;
+
+
+public class CountryInCountry extends Country {
+
+    private Country country;
+
+    public CountryInCountry(Long id, String iso2Code, Country country) {
+        super(id, iso2Code);
+        this.country = country;
+    }
+
+    public Country getCountry() {
+        return country;
+    }
+
+    public void setCountry(Country country) {
+        this.country = country;
+    }
+
+}
