+++ bean-consistency-checker/src/test/resources/log4j.properties
