+++ facade/src/main/java/com/redhat/lightblue/migrator/facade/EntityIdStore.java
@@ -1,0 +1,16 @@
+package com.redhat.lightblue.migrator.facade;
+
+/**
+ * A FIFO queue. Used to pass IDs to create apis in Lightblue DAO without changing the signatures.
+ *
+ *
+ * @author mpatercz
+ *
+ */
+public interface EntityIdStore {
+
+    public void push(Long id);
+
+    public Long pop();
+
+}
