+++ facade/src/test/java/com/redhat/lightblue/migrator/facade/CountryDAOLightblue.java
@@ -1,0 +1,9 @@
+package com.redhat.lightblue.migrator.facade;
+
+import com.redhat.lightblue.migrator.facade.EntityIdStore;
+
+public interface CountryDAOLightblue extends CountryDAO {
+
+    public abstract void setEntityIdStore(EntityIdStore store);
+
+}
