--- utils/src/test/java/com/redhat/lightblue/migrator/utils/CountryDAO.java
@@ -1,13 +1,0 @@
-package com.redhat.lightblue.migrator.utils;
-
-public interface CountryDAO {
-
-    public abstract Country createCountry(Country country);
-
-    public abstract Country createCountryIfNotExists(Country country);
-
-    public abstract Country updateCountry(Country country);
-
-    public abstract Country getCountry(String iso2Code);
-
-}
