--- utils/src/test/java/com/redhat/lightblue/migrator/utils/CountryDAOLightblue.java
@@ -1,7 +1,0 @@
-package com.redhat.lightblue.migrator.utils;
-
-public interface CountryDAOLightblue extends CountryDAO {
-
-    public abstract void setEntityIdStore(EntityIdStore store);
-
-}
