--- utils/src/test/java/com/redhat/lightblue/migrator/utils/consistency/CountryInCountry.java
@@ -1,22 +1,0 @@
-package com.redhat.lightblue.migrator.utils.consistency;
-
-import com.redhat.lightblue.migrator.utils.Country;
-
-public class CountryInCountry extends Country {
-
-    private Country country;
-
-    public CountryInCountry(Long id, String iso2Code, Country country) {
-        super(id, iso2Code);
-        this.country = country;
-    }
-
-    public Country getCountry() {
-        return country;
-    }
-
-    public void setCountry(Country country) {
-        this.country = country;
-    }
-
-}
