--- utils/src/test/java/com/redhat/lightblue/migrator/utils/consistency/ExtendedCountry.java
@@ -1,32 +1,0 @@
-package com.redhat.lightblue.migrator.utils.consistency;
-
-import com.redhat.lightblue.migrator.utils.Country;
-
-public class ExtendedCountry extends Country {
-
-    public ExtendedCountry(Long id, String iso2Code) {
-        super(id, iso2Code);
-    }
-
-    private String additionalFieldConsistencyReq;
-
-    @ConsistencyCheck(ignore=true)
-    private String additionalFieldConsistencyNotReq;
-
-    public String getAdditionalFieldConsistencyReq() {
-        return additionalFieldConsistencyReq;
-    }
-
-    public void setAdditionalFieldConsistencyReq(String additionalFieldConsistencyReq) {
-        this.additionalFieldConsistencyReq = additionalFieldConsistencyReq;
-    }
-
-    public String getAdditionalFieldConsistencyNotReq() {
-        return additionalFieldConsistencyNotReq;
-    }
-
-    public void setAdditionalFieldConsistencyNotReq(String additionalFieldConsistencyNotReq) {
-        this.additionalFieldConsistencyNotReq = additionalFieldConsistencyNotReq;
-    }
-
-}
