--- utils/src/test/java/com/redhat/lightblue/migrator/utils/consistency/VeryExtendedCountry.java
@@ -1,12 +1,0 @@
-package com.redhat.lightblue.migrator.utils.consistency;
-
-public class VeryExtendedCountry extends ExtendedCountry {
-
-    public VeryExtendedCountry(Long id, String iso2Code, String inaccessibleFieldConsistencyReq) {
-        super(id, iso2Code);
-        this.inaccessibleFieldConsistencyReq = inaccessibleFieldConsistencyReq;
-    }
-
-    private String inaccessibleFieldConsistencyReq;
-
-}
